requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pdfminer.six==20250506
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
